solves few errors in all wp

+++ work_package_3/draw.cpp
@@ -138,7 +138,7 @@ void ReadJPEG(char *filename,unsigned char **image,int *width, int *height)
   while (cinfo.output_scanline < cinfo.output_height) {
     jpeg_read_scanlines(&cinfo, buffer, 1);
 
-    for(j=0;j<cinfo.output_width*cinfo.output_components;j++)
+    for(j=0;j<(int)cinfo.output_width*cinfo.output_components;j++)
       {
 	(*image)[i]=buffer[0][j];
 	i++;
